correacao MVC
