--- src/main/java/com/invest7/controller/UserReadController.java
@@ -1,16 +1,0 @@
-package com.invest7.controller;
-
-import com.invest7.dao.UserReadDAO;
-import com.invest7.model.UserRead;
-
-public class UserReadController {
-        public UserRead  lerUser(){
-            int userId = UserSession.getLoggedInUserId();
-            UserReadDAO userDAO= new UserReadDAO();
-            UserRead user = userDAO.readUser(userId);
-            return user;
-
-
-
-    }
-}
