--- src/main/java/com/invest7/model/UserDelete.java
@@ -1,27 +1,0 @@
-package com.invest7.model;
-
-public class UserDelete {
-    private String email;
-    private String senha;
-
-    public UserDelete(String senha, String email) {
-        this.senha = senha;
-        this.email = email;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getSenha() {
-        return senha;
-    }
-
-    public void setSenha(String senha) {
-        this.senha = senha;
-    }
-}
