--- src/main/java/com/invest7/model/UserRead.java
@@ -1,74 +1,0 @@
-package com.invest7.model;
-
-
-import java.util.Date;
-
-public class UserRead {
-    private String nome = null, cpf = null, endereco = null, genero = null, email = null, senhaHash = null;
-    Date dt_nasc;
-
-    public UserRead(String nome, String email, String cpf, String genero, String endereco, Date dt_nasc) {
-        this.nome = nome;
-        this.cpf = cpf;
-        this.genero = genero;
-        this.endereco = endereco;
-        this.email = email;
-        this.dt_nasc = dt_nasc;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public Date getDt_nasc() {
-        return dt_nasc;
-    }
-
-    public void setDt_nasc(Date dt_nasc) {
-        this.dt_nasc = dt_nasc;
-    }
-
-    public String getSenhaHash() {
-        return senhaHash;
-    }
-
-    public void setSenhaHash(String senhaHash) {
-        this.senhaHash = senhaHash;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getGenero() {
-        return genero;
-    }
-
-    public void setGenero(String genero) {
-        this.genero = genero;
-    }
-
-    public String getEndereco() {
-        return endereco;
-    }
-
-    public void setEndereco(String endereco) {
-        this.endereco = endereco;
-    }
-
-    public String getCpf() {
-        return cpf;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-}
