+++ src/main/java/com/invest7/view/CadastroView.java
@@ -1,14 +1,11 @@
 package com.invest7.view;
 
-import com.invest7.controller.DataValidate;
+import com.invest7.controller.*;
 
 import java.util.Date;
 import java.util.Scanner;
 import java.text.SimpleDateFormat;
 import java.text.ParseException;
-import com.invest7.controller.CpfValidate;
-import com.invest7.controller.HashSenha;
-import com.invest7.controller.UserCreateController;
 
 
 public class CadastroView {
@@ -21,7 +18,7 @@ public class CadastroView {
         int opcaoGenero = 0;
         boolean digitoCerto = false;
         MenuPrincipal menuPrincipal = new MenuPrincipal();
-        UserCreateController userCreate = new UserCreateController();
+        UserController userC = new UserController();
         SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy/MM/dd");
 
         System.out.println("-----------TELA DE CADASTRO--------");
@@ -47,7 +44,7 @@ public class CadastroView {
             String cpfValidado = CpfValidate.validaCpf(cpf);
 
             if (cpfValidado != null) {
-                boolean cpfExiste = userCreate.verificaCPF(cpfValidado);
+                boolean cpfExiste = userC.verificaCPF(cpfValidado);
 
                 if (cpfExiste) {
                     System.out.println("CPF já cadastrado. Tente novamente.");
@@ -125,7 +122,7 @@ public class CadastroView {
             System.out.println("6- Digite o seu email: ");
             email = sc.nextLine();
             if (validarEmail(email)) {
-                boolean emailExiste = userCreate.verificaEmail(email);
+                boolean emailExiste = userC.verificaEmail(email);
                 if (emailExiste) {
                     System.out.println("Email já cadastrado. Tente novamente.");
                 } else {
@@ -153,7 +150,7 @@ public class CadastroView {
         senhaHash = senhacrip.hashSenha(senha); //passa a senha cadastrada e transforma em hash
 
 
-        boolean    resultadoCliente = userCreate.criarUser(nome, email, senhaHash, cpf, endereco,genero, data_nasc);
+        boolean    resultadoCliente = userC.criarUser(nome, email, senhaHash, cpf, endereco,genero, data_nasc);
         if (!resultadoCliente) {
             System.out.println("Por algum motivo deu erro, faça o cadastro novamente!!");
         } else {
